# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: ProofQL API
  version: "1.0.0"
  summary: Semantic search over a business's reviews, served at the edge.
  description: |
    ProofQL turns a business's reviews into a semantically searchable corpus so
    a website can show the reviews relevant to each page. This document covers
    the public `/v1` API: pushing reviews in, managing them, and querying them.

    ## Conventions

    - JSON everywhere (`Content-Type: application/json`). Field names are
      `snake_case`. Timestamps are ISO 8601 / RFC 3339 in UTC. Ids are UUIDs.
      A request body with any other `Content-Type` (or none) is a `415
      unsupported_media_type`, refused before it is read.
    - Versioned by path prefix (`/v1`).
    - Every response carries an `x-request-id` header. Quote it in a support
      request; it is also inside every error envelope as `error.request_id`.
      Send your own (up to 128 characters of `A-Z a-z 0-9 . _ : -`) to
      correlate with your logs; anything else is replaced with a fresh id.
    - Every response also carries `X-Content-Type-Options: nosniff`,
      `Referrer-Policy: no-referrer`, and `Cache-Control: no-store` — except
      a successful `GET /v1/query`, which is `Cache-Control: private,
      max-age=0, must-revalidate`: it varies by key, so a shared cache must
      not hold it; the result cache is server-side (`x-cache`).
    - Unknown fields anywhere — body or query string — are a `422
...

    ## Errors

...

    ```json
...
    ```

...

...

...

...

...

...

...

...
servers:
...
security:
  - secretKey: []
paths:
...
  /v1/reviews:
    post:
      tags: [Reviews]
      operationId: ingestReviews
      summary: Push one review or a batch
      description: |
        The push API. Body is one review object or an array of 1 to 100.
        Upsert keyed on `(project, environment, source, external_id)`, where
        project and environment come from the key, never the body:

        - new review → inserted with `status: "indexing"` and queued for the
          pipeline (chunk, embed, sentiment); queryable within seconds.
        - existing, `text` changed (or `rating` removed) → updated and
          re-indexed (`status` back to `"indexing"`).
        - existing, `text` unchanged → the other fields are updated; `status`
          is unchanged.

        Repeats of one `(source, external_id)` inside a batch collapse to the
        last occurrence. The response lists the stored rows in request order.

        The plan's review cap (free: 5,000 per project) is checked before
        anything is written, so a batch either lands whole or not at all
        (`422 review_limit_reached`). The request body may be at most 1 MiB
        (`413 payload_too_large`); 100 maximal reviews fit comfortably.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewIngestBody"
            examples:
              single:
                summary: One review (scope.md §3 "Ingest")
                value:
                  external_id: accounts/1/locations/2/reviews/abc
                  source: google
                  rating: 5
                  text: "Dr. Patel did my implant and I honestly forgot it wasn't my own tooth within a week."
                  author_name: Marcus T.
                  author_avatar_url: null
                  occurred_at: "2026-03-14T18:20:00Z"
                  url: "https://maps.google.com/?cid=123"
                  language: en
                  metadata: { location: north }
              batch:
                summary: A batch of two
                value:
                  - external_id: r-1001
                    source: custom
                    rating: 5
                    text: Painless cleaning, very gentle hygienist.
                    author_name: Dana K.
                    occurred_at: "2026-02-01T09:00:00Z"
                  - external_id: r-1002
                    source: custom
                    rating: null
                    text: Parking behind the building was easy.
                    author_name: J. Alvarez
                    occurred_at: "2026-01-15T17:45:00+01:00"
                    metadata: { location: south }
      responses:
        "200":
          description: |
            Every review in the request was stored (inserted or updated), in
            request order. `status` is `indexing` until the pipeline has
            embedded it, then `indexed`.
          headers:
            x-request-id:
              $ref: "#/components/headers/XRequestId"
            RateLimit-Policy:
              $ref: "#/components/headers/RateLimitPolicy"
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
              examples:
                single:
                  value:
                    reviews:
                      - id: 2f1c9e5a-3b7d-4c1e-9f0a-6d2b8e4a1c3f
                        external_id: accounts/1/locations/2/reviews/abc
                        source: google
                        status: indexing
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/IngestUnprocessable"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    get:
      tags: [Reviews]
      operationId: listReviews
      summary: List reviews
      description: |
        The key's project and environment, newest first, with the filters the
        dashboard's review browser needs. Hidden reviews are included by
        default (`hidden=all`) — this is a management call.

        **Pagination** is keyset, not offset: rows are ordered by
        `(occurred_at DESC NULLS LAST, id DESC)` and each page carries a
        `next_cursor` for the position after its last row. Pass it back as
        `?cursor=` (with the same filters) for the next page; `null` means the
        last page. The cursor is opaque — its encoding may change — and a
        tampered or truncated one is a `422 validation_failed`, never a server
        error. A page is stable while rows are inserted or deleted ahead of it.
      parameters:
        - $ref: "#/components/parameters/ListCursor"
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListSource"
        - $ref: "#/components/parameters/ListMinRating"
        - $ref: "#/components/parameters/ListHidden"
        - $ref: "#/components/parameters/ListSince"
        - $ref: "#/components/parameters/ListIndexed"
      responses:
        "200":
          description: One page.
          headers:
            x-request-id:
              $ref: "#/components/headers/XRequestId"
            RateLimit-Policy:
              $ref: "#/components/headers/RateLimitPolicy"
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListReviewsResponse"
              examples:
                page:
                  value:
                    reviews:
                      - id: 2f1c9e5a-3b7d-4c1e-9f0a-6d2b8e4a1c3f
                        external_id: accounts/1/locations/2/reviews/abc
                        source: google
                        rating: 5
                        text: "Dr. Patel did my implant and I honestly forgot it wasn't my own tooth within a week."
                        author_name: Marcus T.
                        author_avatar_url: null
                        occurred_at: "2026-03-14T18:20:00.000Z"
                        url: "https://maps.google.com/?cid=123"
                        language: en
                        metadata: { location: north }
                        sentiment: positive
                        sentiment_source: rating
                        hidden: false
                        status: indexed
                        created_at: "2026-03-14T18:21:03.412Z"
                        updated_at: "2026-03-14T18:21:03.412Z"
                    next_cursor: eyJvIjoiMjAyNi0wMy0xNFQxODoyMDowMC4wMDBaIiwiaSI6IjJmMWM5ZTVhLTNiN2QtNGMxZS05ZjBhLTZkMmI4ZTRhMWMzZiJ9
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/reviews/{id}:
    parameters:
      - $ref: "#/components/parameters/ReviewId"
    get:
      tags: [Reviews]
      operationId: getReview
      summary: Fetch one review
      responses:
        "200":
          description: The review.
          headers:
            x-request-id:
              $ref: "#/components/headers/XRequestId"
            RateLimit-Policy:
              $ref: "#/components/headers/RateLimitPolicy"
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    patch:
      tags: [Reviews]
      operationId: updateReview
      summary: Hide, unhide, or replace metadata
      description: |
        `hidden: true` removes the review from every query result (sets
        `hidden_at`); `hidden: false` restores it. `metadata` **replaces** the
        whole map. At least one of the two must be present. A body that
        already describes the row changes nothing (no `updated_at` churn).
        Any change purges the project's query cache. Body limit 64 KiB
        (`413 payload_too_large`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewPatch"
            examples:
              hide:
                value: { hidden: true }
              metadata:
                value: { metadata: { location: north, campaign: spring } }
      responses:
        "200":
          description: The review after the change.
          headers:
            x-request-id:
              $ref: "#/components/headers/XRequestId"
            RateLimit-Policy:
              $ref: "#/components/headers/RateLimitPolicy"
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    delete:
      tags: [Reviews]
      operationId: deleteReview
      summary: Delete a review
      description: |
        Removes the review and its excerpts and purges the project's query
        cache. Not idempotent at the HTTP level: a second call is a `404`.
      responses:
        "204":
          description: Deleted. No body.
          headers:
            x-request-id:
              $ref: "#/components/headers/XRequestId"
            RateLimit-Policy:
              $ref: "#/components/headers/RateLimitPolicy"
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimitLimit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    secretKey:
      type: http
      scheme: bearer
      bearerFormat: pq_sk_{live|test}_<32 base62 chars>
      description: |
        A secret key, `Authorization: Bearer pq_sk_live_…`. Servers only.
        Grants everything: ingest, manage, query.
    publishableKey:
      type: http
      scheme: bearer
      bearerFormat: pq_pk_{live|test}_<32 base62 chars>
      description: |
        A publishable key in the `Authorization` header. Query only; the
        request must carry an `Origin` from the project's allowed origins.
    publishableKeyQuery:
      type: apiKey
      in: query
      name: key
      description: |
        A publishable key in the URL (`?key=pq_pk_live_…`) — the snippet's
        form, so the browser's `GET` needs no custom header and no preflight.
        `GET /v1/query` only: on `POST /v1/query` any `?key=` is a `401`
        pointing at the `Authorization` header. A secret key here is a `401`.
        On GET the `Authorization` header wins when both are present.
  parameters:
    ReviewId:
      name: id
      in: path
      required: true
      description: The review's id. A value that is not a UUID is a `404`.
      schema:
        type: string
        format: uuid
      example: 2f1c9e5a-3b7d-4c1e-9f0a-6d2b8e4a1c3f
    ListCursor:
      name: cursor
      in: query
      description: "`next_cursor` from the previous page. Opaque; invalid → `422`."
      schema:
        type: string
        minLength: 1
        pattern: "^[A-Za-z0-9_-]+$"
    ListLimit:
      name: limit
      in: query
      description: Page size.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    ListSource:
      name: source
      in: query
      schema:
        $ref: "#/components/schemas/ReviewSource"
    ListMinRating:
      name: min_rating
      in: query
      description: "`rating >= min_rating`. Unrated reviews are excluded when set."
      schema:
        type: integer
        minimum: 1
        maximum: 5
      example: 4
    ListHidden:
      name: hidden
      in: query
      description: "`true` only hidden, `false` only visible, `all` (default) both."
      schema:
        type: string
        enum: ["true", "false", "all"]
        default: all
    ListSince:
      name: since
      in: query
      description: "`occurred_at >= since`; a full ISO 8601 datetime or a bare `YYYY-MM-DD`."
      schema:
        $ref: "#/components/schemas/DateOrDateTime"
      example: "2025-01-01"
    ListIndexed:
      name: indexed
      in: query
      description: "`true` only indexed (queryable) reviews, `false` only those still indexing."
      schema:
        type: string
        enum: ["true", "false"]
...
  headers:
    XRequestId:
      description: |
        This request's id, on every response. Reuses the caller's
        `x-request-id` when sent (up to 128 chars), else Cloudflare's ray id,
        else a fresh UUID. Also inside every error envelope.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    RateLimitPolicy:
      description: |
        The key's rate-limit policy as `<limit>;w=<window seconds>` (IETF
        draft ratelimit headers), e.g. `300;w=60` for a secret key,
        `120;w=60` for a publishable key. Present once the key is known.
      required: true
      schema:
        type: string
        pattern: "^[0-9]+;w=[0-9]+$"
    RateLimitLimit:
      description: The request limit per window for this key's kind.
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
...
    Unauthorized:
      description: |
        `unauthorized`: no `Authorization` header (and no `?key=` where that
        is allowed), a header that is not `Bearer <key>`, a key that does not
        look like `pq_(sk|pk)_(live|test)_…`, a secret key in the URL, a
        `?key=` on `POST /v1/query`, or an unknown or revoked key. Rate-limit
        headers are absent (the key was not resolved).
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnauthorizedError"
          examples:
            missing:
              value:
                error:
                  code: unauthorized
                  message: "Missing Authorization header. Send `Authorization: Bearer <api key>`."
                  doc_url: https://docs.proofql.com/errors#unauthorized
                  request_id: 1b7c3d9e-2f4a-4b6c-8d0e-1f2a3b4c5d6e
    Forbidden:
      description: |
        `forbidden`: a publishable key (`pq_pk_…`) on a route that requires a
        secret key. Decided before the key is looked up, so rate-limit
        headers are absent.
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ForbiddenError"
...
    PayloadTooLarge:
      description: |
        `payload_too_large`: the body exceeds the route's limit (1 MiB for
        `POST /v1/reviews`, 64 KiB for `PATCH`, 16 KiB for `POST
        /v1/query`). Checked before authentication, so rate-limit headers
        are absent.
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PayloadTooLargeError"
    UnsupportedMediaType:
      description: |
        `unsupported_media_type`: the request has a body whose
        `Content-Type` is not `application/json` (or has none). Refused
        before the body is read and before authentication, so rate-limit
        headers are absent. A request with no body at all (an empty `POST
        /v1/query` asking for the newest reviews) needs no `Content-Type`.
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnsupportedMediaTypeError"
          example:
            error:
              code: unsupported_media_type
              message: "Request bodies must be JSON: send `Content-Type: application/json`."
              doc_url: https://docs.proofql.com/errors#unsupported_media_type
              request_id: 1b7c3d9e-2f4a-4b6c-8d0e-1f2a3b4c5d6e
    ValidationFailed:
      description: |
        `validation_failed`: the body or query string is not valid JSON, does
        not match the schema, or carries an unknown field. `details` lists
        every issue with a dotted `path` (`"0.rating"`, `"filters.limt"`;
        `""` or `"(body)"` for the body as a whole).
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        RateLimit-Policy:
          $ref: "#/components/headers/RateLimitPolicy"
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
          examples:
            unknownField:
              value:
                error:
                  code: validation_failed
                  message: "Invalid request: limt is not a recognized field."
                  doc_url: https://docs.proofql.com/errors#validation_failed
                  request_id: 1b7c3d9e-2f4a-4b6c-8d0e-1f2a3b4c5d6e
                  details:
                    - path: limt
                      message: is not a recognized field
    IngestUnprocessable:
      description: |
        `validation_failed` (see the other routes), or `review_limit_reached`:
        the batch would push the project past its plan's review cap. Nothing
        was written; the message says how many remain.
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        RateLimit-Policy:
          $ref: "#/components/headers/RateLimitPolicy"
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/IngestUnprocessableError"
    RateLimited:
      description: "`rate_limited`: over this key's per-minute limit. Wait `Retry-After` seconds."
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        RateLimit-Policy:
          $ref: "#/components/headers/RateLimitPolicy"
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitedError"
...
    ServiceUnavailable:
      description: |
        `service_unavailable`: the database refused or timed out a connection
        (a connection storm, a paused compute; `Retry-After: 1`), or a daily
        platform allowance is spent (`Retry-After` is the seconds until it
        renews, at most a day). Nothing about the request was wrong; retry
        after `Retry-After` seconds. The message carries the request id and
        nothing about the cause.
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceUnavailableError"
...
    InternalError:
      description: "`internal`: something unexpected. The message carries the request id and nothing about the cause."
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InternalError"
  schemas:
...
    ReviewIngestBody:
      description: One review, or an array of 1 to 100.
      oneOf:
        - $ref: "#/components/schemas/ReviewInput"
        - type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/ReviewInput"
...
    IngestResponse:
      type: object
      required: [reviews]
      additionalProperties: false
      properties:
        reviews:
          type: array
          description: One entry per review in the request, in request order.
          items:
            $ref: "#/components/schemas/IngestedReview"
...
    ListReviewsResponse:
      type: object
      required: [reviews, next_cursor]
      additionalProperties: false
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewResource"
        next_cursor:
          type: [string, "null"]
          description: Pass back as `?cursor=` for the next page; `null` on the last page.
...